First attempt at built-in presets in the CLI.
Works like so:
-Z "Preset Name"
or
--preset "Preset Name"

See a list of all presets and their CLI equivalents (useful for wrappers) this way:
-z
or
--preset-list

Any commands the user adds in addition to in and out locations override that part of the preset. File name extensions are ignored, so if you title the output movie.mp4 but use the Film preset, it will come out as an mkv with an .mp4 extension.

I've also gone and moved where x264opts get duplicated for a 2nd pass, to work better with this system.