fix an off-by-one error in assignment of audio stream registration descriptor format.  This sometimes caused the misdetection of TrueHD audio since it would be tagged as the format of the stream following it in the PMT.
Also fix a problem in PMT parsing that I stumbled upon while investigating
the above problem.  If a PMT has a PID listed more than once for some reason,
we would add that PID multiple times to our stream list.  And if it happened
to be a video PID that is duplicated, the duplications would be interpreted
as audio PIDs.