LinGui: allow multiple instances of the gui to run
Each instance has a queue named "queue.<pid>".  On startup, the gui
opens and locks with lockf a file "ghb.pid.<pid>".  Then it searches
for any other ghb.pid.<pid> files that exist that are not locked.  If
it finds one, then some instance of ghb exited and may have left behind
a queue.  Try to reload it.  If there are no items in the queue, continue
looking for unlocked ghb.pid.<pid> files.