MacGui: Add flexible dylib path manipulation so that we can handle external dylibs more easily At startup, add any extra dylib paths to DYLD_FALLBACK_LIBRARY_PATH.  This is the last path searched by the system for dylibs and we add our paths to the end of it's list, so this will never override any system libs or paths the user has explicitly set.
Since applications read the environment once at startup, these changes don't take effect
until we restart with execv.  In order to avoid a possible exec bomb, we add a parameter to
the argv list to prevent any further exec's.

Note that this causes a minor glitch when running under gdb. The execv triggers a trap.  You
just have to 'continue'.