Improved logging of keyint min/max And small tweak to keyint_max calculation.
x264 slightly changed the way the automatic min-keyint is calculated. 
Also, keyint infinite was added. This does not cause any encoding issues, 
but depending on the settings and the framerate, the log contained
keyint values that don't correspond to the actual encode's values.

The keyint_max was calculated, 59.94 fps sources will get a keyint of 599 
rather than 600.  With this change, It is rounded to the closest integer 
then multiplied by 10. 

Thanks to Rodeo